Fix memory leak and double-free in multi-threaded use of InstructionFactory.
The two std::map arrays (encoding_maps and mnemonic_maps) have a static
lifetime in InstructionFactory. Thus, they should be responsible for destroying
their contents as their lifetime is not bound to a particular InstructionFactory
instance (the desctructor newer gets called). This was causing memory leaks in
tests:

==156499== 104 bytes in 1 blocks are definitely lost in loss record 4 of 106
==156499==    at 0x4844ED1: operator new(unsigned long) (in /libexec/valgrind/vgpreload_memcheck-amd64-linux.so)
==156499==    by 0x4951F8D: spect::InstructionFactory::Initialize() (InstructionFactory.cpp:53)
==156499==    by 0x495493B: __static_initialization_and_destruction_0() (InstructionFactory.cpp:113)
==156499==    by 0x4954953: _GLOBAL__sub_I_InstructionFactory.cpp (InstructionFactory.cpp:115)
==156499==    by 0x400508D: call_init (in /lib/ld-linux-x86-64.so.2)
==156499==    by 0x400517B: _dl_init (in /lib/ld-linux-x86-64.so.2)
==156499==    by 0x401BC2F: ??? (in /lib/ld-linux-x86-64.so.2)

This is fixed by using shared_ptrs and by getting rid of the destructor trying to
delete the individual instructions.

Furthermore, the static lifetime was causing havoc in multi-threaded use,
this was fixed by making the arrays to be thread_local.